wine64: Sync IMAGE_DOS_HEADER with Wine headers to fix 64-bit issues.
https://bugs.kde.org/show_bug.cgi?id=253657

+++ coregrind/m_debuginfo/readpdb.c
@@ -109,25 +109,25 @@ typedef  UChar  BYTE;
 
 #pragma pack(2)
 typedef struct _IMAGE_DOS_HEADER {
-    unsigned short  e_magic;      /* 00: MZ Header signature */
-    unsigned short  e_cblp;       /* 02: Bytes on last page of file */
-    unsigned short  e_cp;         /* 04: Pages in file */
-    unsigned short  e_crlc;       /* 06: Relocations */
-    unsigned short  e_cparhdr;    /* 08: Size of header in paragraphs */
-    unsigned short  e_minalloc;   /* 0a: Minimum extra paragraphs needed */
-    unsigned short  e_maxalloc;   /* 0c: Maximum extra paragraphs needed */
-    unsigned short  e_ss;         /* 0e: Initial (relative) SS value */
-    unsigned short  e_sp;         /* 10: Initial SP value */
-    unsigned short  e_csum;       /* 12: Checksum */
-    unsigned short  e_ip;         /* 14: Initial IP value */
-    unsigned short  e_cs;         /* 16: Initial (relative) CS value */
-    unsigned short  e_lfarlc;     /* 18: File address of relocation table */
-    unsigned short  e_ovno;       /* 1a: Overlay number */
-    unsigned short  e_res[4];     /* 1c: Reserved words */
-    unsigned short  e_oemid;      /* 24: OEM identifier (for e_oeminfo) */
-    unsigned short  e_oeminfo;    /* 26: OEM information; e_oemid specific */
-    unsigned short  e_res2[10];   /* 28: Reserved words */
-    unsigned long   e_lfanew;     /* 3c: Offset to extended header */
+    WORD  e_magic;      /* 00: MZ Header signature */
+    WORD  e_cblp;       /* 02: Bytes on last page of file */
+    WORD  e_cp;         /* 04: Pages in file */
+    WORD  e_crlc;       /* 06: Relocations */
+    WORD  e_cparhdr;    /* 08: Size of header in paragraphs */
+    WORD  e_minalloc;   /* 0a: Minimum extra paragraphs needed */
+    WORD  e_maxalloc;   /* 0c: Maximum extra paragraphs needed */
+    WORD  e_ss;         /* 0e: Initial (relative) SS value */
+    WORD  e_sp;         /* 10: Initial SP value */
+    WORD  e_csum;       /* 12: Checksum */
+    WORD  e_ip;         /* 14: Initial IP value */
+    WORD  e_cs;         /* 16: Initial (relative) CS value */
+    WORD  e_lfarlc;     /* 18: File address of relocation table */
+    WORD  e_ovno;       /* 1a: Overlay number */
+    WORD  e_res[4];     /* 1c: Reserved words */
+    WORD  e_oemid;      /* 24: OEM identifier (for e_oeminfo) */
+    WORD  e_oeminfo;    /* 26: OEM information; e_oemid specific */
+    WORD  e_res2[10];   /* 28: Reserved words */
+    DWORD e_lfanew;     /* 3c: Offset to extended header */
 } IMAGE_DOS_HEADER, *PIMAGE_DOS_HEADER;
 
 #define IMAGE_DOS_SIGNATURE    0x5A4D     /* MZ   */
